+++ io.github.ericmedvet.jgea.core/src/main/java/io/github/ericmedvet/jgea/core/solver/bi/mapelites/MapElitesBiEvolver.java
@@ -31,11 +31,13 @@ import io.github.ericmedvet.jgea.core.solver.mapelites.MEIndividual;
 import io.github.ericmedvet.jgea.core.solver.mapelites.MEPopulationState;
 import io.github.ericmedvet.jgea.core.solver.mapelites.MapElites;
 import io.github.ericmedvet.jgea.core.util.Misc;
+import io.github.ericmedvet.jnb.datastructure.Pair;
 import java.util.*;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.BiFunction;
 import java.util.function.BinaryOperator;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -48,6 +50,8 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
   private final Mutation<G> mutation;
   private final List<MapElites.Descriptor<G, S, Q>> descriptors;
   private final boolean emptyArchive;
+  private final MapElitesBiEvolver.OpponentSelector<G, S, Q, O> opponentsSelector;
+  private final Function<List<Q>, Q> fitnessAggregator;
 
   public MapElitesBiEvolver(
       Function<? super G, ? extends S> solutionMapper,
@@ -58,13 +62,27 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
       List<MapElites.Descriptor<G, S, Q>> descriptors,
       BinaryOperator<Q> fitnessReducer,
       boolean emptyArchive,
-      List<PartialComparator<? super MEIndividual<G, S, Q>>> additionalIndividualComparators
+      List<PartialComparator<? super MEIndividual<G, S, Q>>> additionalIndividualComparators,
+      MapElitesBiEvolver.OpponentSelector<G, S, Q, O> opponentsSelector,
+      Function<List<Q>, Q> fitnessAggregator
   ) {
     super(solutionMapper, genotypeFactory, stopCondition, false, fitnessReducer, additionalIndividualComparators);
     this.populationSize = populationSize;
     this.mutation = mutation;
     this.descriptors = descriptors;
     this.emptyArchive = emptyArchive;
+    this.opponentsSelector = opponentsSelector;
+    this.fitnessAggregator = fitnessAggregator;
+  }
+
+  @FunctionalInterface
+  public interface OpponentSelector<G, S, Q, O> {
+    List<MEIndividual<G, S, Q>> select(
+        Collection<MEIndividual<G, S, Q>> population,
+        MEIndividual<G, S, Q> individual,
+        QualityBasedBiProblem<S, O, Q> problem,
+        RandomGenerator random
+    );
   }
 
   @Override
@@ -75,57 +93,50 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
   ) throws SolverException {
     // create new genotypes and split in two list of opponents
     AtomicLong counter = new AtomicLong(0);
-    List<? extends G> genotypes = genotypeFactory.build(populationSize, random);
-    List<? extends G> firstHalfGenotypes;
-    int half = populationSize / 2;
-    if (populationSize % 2 == 0) {
-      firstHalfGenotypes = genotypes.subList(0, half);
-    } else {
-      firstHalfGenotypes = genotypes.subList(0, half + 1);
+    Collection<? extends G> genotypes = genotypeFactory.build(populationSize, random);
+    Collection<MEIndividual<G, S, Q>> individuals = genotypes.stream()
+        .map(
+            g -> MEIndividual.from(
+                Individual.of(
+                    counter.getAndIncrement(),
+                    g,
+                    solutionMapper.apply(g),
+                    null,
+                    0,
+                    0,
+                    List.of()
+                ),
+                descriptors
+            )
+        )
+        .toList();
+    Map<Pair<Long, Long>, Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>>> fights = new HashMap<>();
+    BiFunction<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>, Pair<Long, Long>> individualsToIdPair = (
+        i1,
+        i2
+    ) -> new Pair<>(Math.min(i1.id(), i2.id()), Math.max(i1.id(), i2.id()));
+    for (MEIndividual<G, S, Q> individual : individuals) {
+      List<MEIndividual<G, S, Q>> opponents = opponentsSelector.select(individuals, individual, problem, random);
+      for (MEIndividual<G, S, Q> opponent : opponents) {
+        fights.putIfAbsent(individualsToIdPair.apply(individual, opponent), new Pair<>(individual, opponent));
+      }
     }
-    List<? extends G> secondHalfGenotypes = genotypes.subList(half, populationSize);
-    // build futures
-    List<Future<List<MEIndividual<G, S, Q>>>> futures = new ArrayList<>();
-    for (int i = 0; i < firstHalfGenotypes.size(); i++) {
+    List<Future<FightOutcome<Q>>> futures = new ArrayList<>();
+    List<Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>>> fightsOpponents = fights.values().stream().toList();
+    for (int i = 0; i < fights.size(); i++) {
       int index = i;
-      Future<List<MEIndividual<G, S, Q>>> future = executor.submit(() -> {
-        G firstGenotype = firstHalfGenotypes.get(index);
-        G secondGenotype = secondHalfGenotypes.get(index);
-        S firstSolution = solutionMapper.apply(firstGenotype);
-        S secondSolution = solutionMapper.apply(secondGenotype);
-        O outcome = problem.outcomeFunction().apply(firstSolution, secondSolution);
+      Future<FightOutcome<Q>> future = executor.submit(() -> {
+        Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>> fight = fightsOpponents.get(index);
+        MEIndividual<G, S, Q> i1 = fight.first();
+        MEIndividual<G, S, Q> i2 = fight.second();
+        O outcome = problem.outcomeFunction().apply(i1.solution(), i2.solution());
         Q firstQuality = problem.firstQualityFunction().apply(outcome);
         Q secondQuality = problem.secondQualityFunction().apply(outcome);
-        MEIndividual<G, S, Q> firstIndividual = MEIndividual.from(
-            Individual.of(
-                counter.getAndIncrement(),
-                firstGenotype,
-                firstSolution,
-                firstQuality,
-                0,
-                0,
-                List.of()
-            ),
-            descriptors
-        );
-        MEIndividual<G, S, Q> secondIndividual = MEIndividual.from(
-            Individual.of(
-                counter.getAndIncrement(),
-                secondGenotype,
-                secondSolution,
-                secondQuality,
-                0,
-                0,
-                List.of()
-            ),
-            descriptors
-        );
-        return List.of(firstIndividual, secondIndividual);
+        return new FightOutcome<>(i1.id(), i2.id(), firstQuality, secondQuality);
       });
       futures.add(future);
     }
-    // extract future results
-    Collection<MEIndividual<G, S, Q>> individuals = futures.stream()
+    Collection<FightOutcome<Q>> fightsOutcomes = futures.stream()
         .map(listFuture -> {
           try {
             return listFuture.get();
@@ -133,7 +144,16 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
             throw new RuntimeException(e);
           }
         })
-        .flatMap(Collection::stream)
+        .toList();
+    Map<Long, List<Q>> idsFitnessMap = new HashMap<>();
+    fightsOutcomes.forEach(fo -> {
+      idsFitnessMap.computeIfAbsent(fo.id1(), k -> new ArrayList<>()).add(fo.f1());
+      idsFitnessMap.computeIfAbsent(fo.id2(), k -> new ArrayList<>()).add(fo.f2());
+    });
+    Collection<MEIndividual<G, S, Q>> updatedIndividuals = individuals.stream()
+        .map(
+            i -> i.updateQuality(fitnessAggregator.apply(idsFitnessMap.get(i.id())), 0)
+        )
         .toList();
     MEPopulationState<G, S, Q, QualityBasedBiProblem<S, O, Q>> newState = MEPopulationState.empty(
         problem,
@@ -143,7 +163,7 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
     return newState.updatedWithIteration(
         populationSize,
         futures.size(),
-        newState.archive().updated(individuals, MEIndividual::bins, partialComparator(problem))
+        newState.archive().updated(updatedIndividuals, MEIndividual::bins, partialComparator(problem))
     );
   }
 
@@ -154,9 +174,9 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
       MEPopulationState<G, S, Q, QualityBasedBiProblem<S, O, Q>> state
   ) throws SolverException {
     // build offspring with empty quality
-    List<MEIndividual<G, S, Q>> individuals = new ArrayList<>(state.archive().asMap().values().stream().toList());
+    Collection<MEIndividual<G, S, Q>> individuals = new ArrayList<>(state.archive().asMap().values().stream().toList());
     AtomicLong counter = new AtomicLong(state.nOfBirths());
-    List<ChildGenotype<G>> newChildGenotypes = IntStream.range(0, populationSize)
+    Collection<ChildGenotype<G>> newChildGenotypes = IntStream.range(0, populationSize)
         .mapToObj(j -> Misc.pickRandomly(individuals, random))
         .map(
             p -> new ChildGenotype<>(
@@ -176,53 +196,41 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
             )
             .toList()
     );
-    // randomly "pair" individuals
-    Collections.shuffle(individuals, random);
-    List<MEIndividual<G, S, Q>> firstHalfIndividuals;
-    int half = individuals.size() / 2;
-    if (individuals.size() % 2 == 0) {
-      firstHalfIndividuals = individuals.subList(0, half);
-    } else {
-      firstHalfIndividuals = individuals.subList(0, half + 1);
+    Map<Pair<Long, Long>, Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>>> fights = new HashMap<>();
+    BiFunction<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>, Pair<Long, Long>> individualsToIdPair = (
+        i1,
+        i2
+    ) -> new Pair<>(
+        Math.min(i1.id(), i2.id()),
+        Math.max(i1.id(), i2.id())
+    );
+    for (MEIndividual<G, S, Q> individual : individuals) {
+      List<MEIndividual<G, S, Q>> opponents = opponentsSelector.select(
+          individuals,
+          individual,
+          state.problem(),
+          random
+      );
+      for (MEIndividual<G, S, Q> opponent : opponents) {
+        fights.putIfAbsent(individualsToIdPair.apply(individual, opponent), new Pair<>(individual, opponent));
+      }
     }
-    List<MEIndividual<G, S, Q>> secondHalfIndividuals = individuals.subList(half, individuals.size());
-    // build futures
-    List<Future<List<MEIndividual<G, S, Q>>>> futures = new ArrayList<>();
-    for (int i = 0; i < firstHalfIndividuals.size(); i++) {
+    List<Future<FightOutcome<Q>>> futures = new ArrayList<>();
+    List<Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>>> fightsOpponents = fights.values().stream().toList();
+    for (int i = 0; i < fights.size(); i++) {
       int index = i;
-      Future<List<MEIndividual<G, S, Q>>> future = executor.submit(() -> {
-        MEIndividual<G, S, Q> firstIndividual = firstHalfIndividuals.get(index);
-        MEIndividual<G, S, Q> secondIndividual = secondHalfIndividuals.get(index);
-        O outcome = state.problem()
-            .outcomeFunction()
-            .apply(firstIndividual.solution(), secondIndividual.solution());
-        Q newFirstQuality = state.problem().firstQualityFunction().apply(outcome);
-        Q newSecondQuality = state.problem().secondQualityFunction().apply(outcome);
-        firstIndividual = Objects.isNull(firstIndividual.quality()) ? firstIndividual.updateQuality(
-            newFirstQuality,
-            state.nOfIterations()
-        ) : firstIndividual.updateQuality(
-            fitnessReducer.apply(firstIndividual.quality(), newFirstQuality),
-            state.nOfIterations()
-        );
-        secondIndividual = Objects.isNull(secondIndividual.quality()) ? secondIndividual.updateQuality(
-            newSecondQuality,
-            state.nOfIterations()
-        ) : secondIndividual.updateQuality(
-            fitnessReducer.apply(secondIndividual.quality(), newSecondQuality),
-            state.nOfIterations()
-        );
-        // If populationSize is odd, the last element of firstHalfGenotypes is the same as the first element of
-        // secondHalfGenotypes
-        if (individuals.size() % 2 != 0 && index != 0) {
-          return List.of(firstIndividual);
-        }
-        return List.of(firstIndividual, secondIndividual);
+      Future<FightOutcome<Q>> future = executor.submit(() -> {
+        Pair<MEIndividual<G, S, Q>, MEIndividual<G, S, Q>> fight = fightsOpponents.get(index);
+        MEIndividual<G, S, Q> i1 = fight.first();
+        MEIndividual<G, S, Q> i2 = fight.second();
+        O outcome = state.problem().outcomeFunction().apply(i1.solution(), i2.solution());
+        Q firstQuality = state.problem().firstQualityFunction().apply(outcome);
+        Q secondQuality = state.problem().secondQualityFunction().apply(outcome);
+        return new FightOutcome<>(i1.id(), i2.id(), firstQuality, secondQuality);
       });
       futures.add(future);
     }
-    // Collect futures
-    Collection<MEIndividual<G, S, Q>> newAndOldPopulation = futures.stream()
+    Collection<FightOutcome<Q>> fightsOutcomes = futures.stream()
         .map(listFuture -> {
           try {
             return listFuture.get();
@@ -230,7 +238,22 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
             throw new RuntimeException(e);
           }
         })
-        .flatMap(Collection::stream)
+        .toList();
+    Map<Long, List<Q>> idsFitnessMap = new HashMap<>();
+    fightsOutcomes.forEach(fo -> {
+      idsFitnessMap.computeIfAbsent(fo.id1(), k -> new ArrayList<>()).add(fo.f1());
+      idsFitnessMap.computeIfAbsent(fo.id2(), k -> new ArrayList<>()).add(fo.f2());
+    });
+    Collection<MEIndividual<G, S, Q>> updatedIndividuals = individuals.stream()
+        .map(
+            i -> {
+              Q newQuality = i.quality() != null ? fitnessReducer.apply(
+                  i.quality(),
+                  fitnessAggregator.apply(idsFitnessMap.get(i.id()))
+              ) : fitnessAggregator.apply(idsFitnessMap.get(i.id()));
+              return i.updateQuality(newQuality, i.qualityMappingIteration());
+            }
+        )
         .toList();
     PartialComparator<? super Individual<?, ?, ?>> updaterComparator = (
         newI,
@@ -240,9 +263,12 @@ public class MapElitesBiEvolver<G, S, Q, O> extends AbstractBiEvolver<MEPopulati
     if (emptyArchive) {
       archive = new Archive<>(state.archive().binUpperBounds());
     } else {
-      archive = state.archive().updated(newAndOldPopulation, MEIndividual::bins, updaterComparator);
+      archive = state.archive().updated(updatedIndividuals, MEIndividual::bins, updaterComparator);
     }
-    archive = archive.updated(newAndOldPopulation, MEIndividual::bins, partialComparator(state.problem()));
+    archive = archive.updated(updatedIndividuals, MEIndividual::bins, partialComparator(state.problem()));
     return state.updatedWithIteration(populationSize, futures.size(), archive);
+  }
+
+  public record FightOutcome<Q>(long id1, long id2, Q f1, Q f2) {
   }
 }
