+++ io.github.ericmedvet.jgea.experimenter/src/main/resources/exp-examples/pong-generalized-bi-me.txt
@@ -19,7 +19,7 @@ ea.experiment(
   runs =
     (randomGenerator = (seed = [1:1:5]) * [m.defaultRG()]) *
     (solver = (nEval = [100000]) * [
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-random1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -32,7 +32,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-random2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -45,7 +45,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-random3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -58,7 +58,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 3);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-random1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -71,7 +71,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-random2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -84,7 +84,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-random3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -97,7 +97,7 @@ ea.experiment(
         opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 3);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-farthest1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -110,7 +110,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-nearest1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -123,7 +123,7 @@ ea.experiment(
         opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-farthest2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -136,7 +136,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-nearest2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -149,7 +149,7 @@ ea.experiment(
         opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-farthest3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -162,7 +162,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 3);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-tree-nearest3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
@@ -175,7 +175,7 @@ ea.experiment(
         opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 3);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-farthest1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -188,7 +188,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-nearest1";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -201,7 +201,7 @@ ea.experiment(
         opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 1);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-farthest2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -214,7 +214,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-nearest2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -227,7 +227,7 @@ ea.experiment(
         opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-farthest3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -240,7 +240,7 @@ ea.experiment(
         opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 3);
         fitnessAggregator = f.avg()
       );
-      ea.s.generalizedBiMapElites(
+      ea.s.biMapElites(
         name = "bi-me-ann-nearest3";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.doubleString();
@@ -292,28 +292,6 @@ ea.experiment(
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
         mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf())
-      );
-      ea.s.biMapElites(
-        name = "bi-me-ann";
-        fitnessReducer = ea.misc.lossyAverage();
-        representation = ea.r.doubleString();
-        mapper = ea.m.dsToNpnds(npnds = ds.num.mlp(innerLayerRatio = 4; nOfInnerLayers = 1));
-        emptyArchive = false;
-        descriptors = [ % Prendere descrittori Lippolis "covering.txt"
-          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 0)); min = -1; max = 1);
-          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 1)); min = -1; max = 1)
-        ]
-      );
-      ea.s.biMapElites(
-        name = "bi-me-tree";
-        fitnessReducer = ea.misc.lossyAverage();
-        representation = ea.r.srTree();
-        mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf());
-        emptyArchive = false;
-        descriptors = [
-          ea.s.me.d.descriptor(f = ea.f.treeSize(of = ea.f.genotype()); min = 0; max = 250; nOfBins = 10);
-          ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.treeLeaves(of = ea.f.genotype()); condition = predicate.matches(f = f.toString(); regex = "i[0-9]+"))); min = 0; max = 30; nOfBins = 10)
-        ]
       )
 ]) *
     (problem = [ea.p.biSimToBqb(
