+++ io.github.ericmedvet.jgea.experimenter/src/main/resources/exp-examples/pong-ss-roarnet.txt
@@ -15,10 +15,41 @@ $VS_OPPONENT = ds.f.opponentBiSimulator(
   opponent = ds.opponent.pong.simple();
   simulation = $SIMULATION
 )
+
+$MLP = ds.num.mlp(innerLayerRatio = 4; nOfInnerLayers = 1)
+
+$THRESHOLD = 0.2
+
 ea.experiment(
   runs =
-    (randomGenerator = (seed = [1:1:3]) * [m.defaultRG()]) *
-    (solver = (nEval = [30000]) * [
+    (randomGenerator = (seed = [1:1:10]) * [m.defaultRG()]) *
+    (solver = (nEval = [50000]) * [
+      ea.s.biMapElites(
+        name = "bi-me-tree-random2";
+        fitnessReducer = ea.misc.lossyAverage();
+        representation = ea.r.srTree();
+        mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf());
+        emptyArchive = false;
+        descriptors = [
+          ea.s.me.d.descriptor(f = ea.f.treeSize(of = ea.f.genotype()); min = 0; max = 250; nOfBins = 10);
+          ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.treeLeaves(of = ea.f.genotype()); condition = predicate.matches(f = f.toString(); regex = "i[0-9]+"))); min = 0; max = 30; nOfBins = 10)
+        ];
+        opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 2);
+        fitnessAggregator = f.avg()
+      );
+      ea.s.biMapElites(
+        name = "bi-me-ann-random2";
+        fitnessReducer = ea.misc.lossyAverage();
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP);
+        emptyArchive = false;
+        descriptors = [
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 0)); min = -1; max = 1);
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 1)); min = -1; max = 1)
+        ];
+        opponentsSelector = ea.misc.randomMESelector(nOfOpponents = 2);
+        fitnessAggregator = f.avg()
+      );
       ea.s.biMapElites(
         name = "bi-me-tree-farthest2";
         fitnessReducer = ea.misc.lossyAverage();
@@ -33,7 +64,7 @@ ea.experiment(
         fitnessAggregator = f.avg()
       );
       ea.s.biMapElites(
-        name = "bi-me-tree-best1";
+        name = "bi-me-tree-nearest2";
         fitnessReducer = ea.misc.lossyAverage();
         representation = ea.r.srTree();
         mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf());
@@ -42,45 +73,83 @@ ea.experiment(
           ea.s.me.d.descriptor(f = ea.f.treeSize(of = ea.f.genotype()); min = 0; max = 250; nOfBins = 10);
           ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.treeLeaves(of = ea.f.genotype()); condition = predicate.matches(f = f.toString(); regex = "i[0-9]+"))); min = 0; max = 30; nOfBins = 10)
         ];
-        opponentsSelector = ea.misc.bestMESelector(nOfOpponents = 1);
+        opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
       ea.s.biMapElites(
-        name = "bi-me-tree-best2";
+        name = "bi-me-ann-farthest2";
         fitnessReducer = ea.misc.lossyAverage();
-        representation = ea.r.srTree();
-        mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf());
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP);
         emptyArchive = false;
         descriptors = [
-          ea.s.me.d.descriptor(f = ea.f.treeSize(of = ea.f.genotype()); min = 0; max = 250; nOfBins = 10);
-          ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.treeLeaves(of = ea.f.genotype()); condition = predicate.matches(f = f.toString(); regex = "i[0-9]+"))); min = 0; max = 30; nOfBins = 10)
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 0)); min = -1; max = 1);
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 1)); min = -1; max = 1)
         ];
-        opponentsSelector = ea.misc.bestMESelector(nOfOpponents = 2);
+        opponentsSelector = ea.misc.farthestMESelector(nOfOpponents = 2);
         fitnessAggregator = f.avg()
       );
       ea.s.biMapElites(
-        name = "bi-me-tree-best3";
+        name = "bi-me-ann-nearest2";
         fitnessReducer = ea.misc.lossyAverage();
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP);
+        emptyArchive = false;
+        descriptors = [
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 0)); min = -1; max = 1);
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 1)); min = -1; max = 1)
+        ];
+        opponentsSelector = ea.misc.nearestMESelector(nOfOpponents = 2);
+        fitnessAggregator = f.avg()
+      );
+      ea.s.ga(
+        name = "self-ga-tree";
+        representation = ea.r.srTree();
+        mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf())
+      );
+      ea.s.ga(
+        name = "self-ga-ann";
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP)
+      );
+      ea.s.mapElites(
+        name = "self-me-ann";
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP);
+        descriptors = [ % Prendere descrittori Lippolis "covering.txt"
+          % ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.getLayerWeights(of = $MLP), condition = predicate.lt(t = $THRESHOLD))));
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 0)); min = -1; max = 1);
+          ea.s.me.d.descriptor(f = f.avg(of = f.nkTh(of = ea.f.genotype(); n = 2; k = 1)); min = -1; max = 1)
+        ]
+      );
+      ea.s.mapElites(
+        name = "self-me-tree";
         representation = ea.r.srTree();
         mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf());
-        emptyArchive = false;
         descriptors = [
           ea.s.me.d.descriptor(f = ea.f.treeSize(of = ea.f.genotype()); min = 0; max = 250; nOfBins = 10);
           ea.s.me.d.descriptor(f = f.size(of = f.filter(of = ea.f.treeLeaves(of = ea.f.genotype()); condition = predicate.matches(f = f.toString(); regex = "i[0-9]+"))); min = 0; max = 30; nOfBins = 10)
-        ];
-        opponentsSelector = ea.misc.bestMESelector(nOfOpponents = 3);
-        fitnessAggregator = f.avg()
+        ]
+      );
+      ea.s.biGa(
+        name = "bi-ga-ann";
+        fitnessReducer = ea.misc.lossyAverage();
+        representation = ea.r.doubleString();
+        mapper = ea.m.dsToNpnds(npnds = $MLP)
+      );
+      ea.s.biGa(
+        name = "bi-ga-tree";
+        fitnessReducer = ea.misc.lossyAverage();
+        representation = ea.r.srTree();
+        mapper = ea.m.nurfToNds(of = ea.m.srTreeToNurf())
       )
 ]) *
-
-    (problem = [
-      ea.p.biSimToBqb(
-        simulation = $SIMULATION;
-        qFunction1 = ds.e.pong.shiftedScoreDiff1();
-        qFunction2 = ds.e.pong.shiftedScoreDiff2();
-        type = maximize
-      )
-    ]) *
+    (problem = [ea.p.biSimToBqb(
+      simulation = $SIMULATION;
+      qFunction1 = ds.e.pong.scoreDiff1();
+      qFunction2 = ds.e.pong.scoreDiff2();
+      type = maximize
+    )]) *
     [ea.run()];
   listeners = [
     ea.l.console(
@@ -130,14 +199,13 @@ ea.experiment(
         video = ea.misc.toVideo(drawer = ds.d.pong());
         of = $VS_OPPONENT
       );
-      path = "../results/{name}/{startTime}/best-vs-simple-{run.solver.name}--{run.randomGenerator.seed:%04d}"
+      path = "../results/{name}/{startTime}/best-vs-simple-{run.solver.name}-{run.randomGenerator.seed:%04d}"
     );
     ea.l.bestCsv(
         path = "../results/{name}/{startTime}/allBest.csv";
         runFunctions = [
           ea.f.runKey(key = "run.problem.simulation.environment");
-          ea.f.runKey(key = "run.solver.mapper");
-          ea.f.runKey(key = "run.problem.qFunction1")
+          ea.f.runKey(key = "run.solver.mapper")
         ];
         functions = [
           f.toBase64(of = ea.f.genotype(of = ea.f.best()))
